SmartBar: reset button animation/pressed state when cancelled
If the notification shade was pulled down and the shade close
swipe started from SmartBar button, the button would not handle
the cancel motion event. This left the button in a pressed
animation state. Also, clear action callbacks as well here. It's
likely the user does not want to fire a action if the motion
event is cancelled.

Change-Id: Ide1e27ed91af330c8cf494f4beb99b5a447e40c8
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>